lvl-61: fix mocks
